White Cat ESP32
Project White Cat – A talisman. 

A Work in progress the cute white cat from oneko a popular open source desktop amusement ported to this Internet of things device. I have no idea whats going to be the final release of this program. I just keep working on it writing as I go. I have this idea that hackers in the near future will be sporting technology or gear that is A talisman of protection & attack. Why not a watch? Or finger sized computer and/or micro-controller. 

Hacking stuff it can do So far:  

ONE: Anonymous LAN proxy that lists local networks. Our ESSID name and password are simple and works just like connecting to a wireless hot spot. Once connected to the white cat ESSID you can use A browser and just enter in the IP address. The software will then return A HTML page with network names and other information. (your basic pen-testing free roaming Wi-Fi station) 

TWO: Panic Modes both buttons A and B can be pressed to activate payloads that hide our ESSID network and flood the Wi-Fi network’s with beacon Spam. One payload is a Alphanumeric attack, The other is A standard list of stupid Wi-Fi names.  (your basic Wi-Fi flood attack)