D1 mini Pro Whitecat port
This project is really quite a buggy outline. Will have to rewrite this code. It's a port of the cat artwork I did for the M5StickC re-engineered for the O-Led black & white display. It works with the 8x8 red led matrix shield. 
  
This is a list of all shields and parts to make this project: 
1. D1 mini pro 16MB flash version. 
2. Data logger shield SD/RTC clock combo. 
3. PIR shield 
4. SHT30 shield 
5. 0.66" O-LED 64x45 shield 
6. 8x8 Matris Shield 
7. Triple base Shield